proposal.go proposal_interface.go serializable_proposal.go
